# site/public/openapi.yaml
openapi: 3.1.0
info:
  title: Cadastrum Public API
  version: 1.2.0
  description: |
    Cadastrum API — gayrimenkul fiyat tahmin, emsal sorgu ve risk verisi.
    Kurumsal Pro tier'a özel programmatic erişim.

    Token formatı: `cdrm_<32-hex>`. /v1/api/tokens endpoint'inden üretilir.
    Rate limit varsayılan: 60 req/dakika/token (oluştururken özelleştirilebilir).

    **Değişiklik geçmişi:**
    - v1.2.0: /fiyat/trend, /fiyat/toplu-ozet, /fiyat/toplu-ilce-ozet, /sorgu endpoint'leri eklendi.
              /fiyat/{il,ilce} endpoint şemaları tamamlandı. Seed endpoint'leri (internal) belgelendi.
    - v1.1.0: İlk public release.
  contact:
    name: Cadastrum
    email: eparlak996@gmail.com
servers:
  - url: https://cadastrum-api.cadastrum-tr.workers.dev
    description: Production (Cloudflare Worker)
  - url: https://api.cadastrum.com.tr
    description: Production (custom domain)
security:
  - ApiKeyAuth: []
paths:
...
  /v1/fiyat/mahalle/{il}/{ilce}/{mahalle}:
    get:
      tags: [Fiyat]
      summary: Mahalle bazlı fiyat istatistiği
      description: |
        Scraper verileriyle hesaplanan mahalle medyanı. Veri yoksa AI baseline fallback döner.
      security: []
      parameters:
        - { name: il, in: path, required: true, schema: { type: string }, example: istanbul }
        - { name: ilce, in: path, required: true, schema: { type: string }, example: besiktas }
        - { name: mahalle, in: path, required: true, schema: { type: string }, example: levent }
        - name: kategori
          in: query
          schema: { type: string, default: arsa, enum: [arsa, tarla, konut, bahce, bag, zeytinlik] }
      responses:
        "200":
          description: Mahalle istatistik
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MahalleIstatistik' }
        "404":
          description: Veri yok
  /v1/fiyat/ilce/{il}/{ilce}:
    get:
      tags: [Fiyat]
      summary: İlçe bazlı fiyat istatistiği
      security: []
      parameters:
        - { name: il, in: path, required: true, schema: { type: string } }
        - { name: ilce, in: path, required: true, schema: { type: string } }
        - { name: kategori, in: query, schema: { type: string, default: arsa, enum: [arsa, tarla, konut, bahce, bag, zeytinlik] } }
      responses:
        "200":
          description: İlçe istatistik + mahalle listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  medyan: { type: number }
                  q1: { type: number }
                  q3: { type: number }
                  ilan_adet: { type: integer }
                  kaynak: { type: string, enum: [ilan-istatistik, ai-aggregate] }
                  mahalleler:
                    type: array
                    items:
                      type: object
                      properties:
                        mahalle_norm: { type: string }
                        medyan: { type: number }
                        ilan_adet: { type: integer }
  /v1/fiyat/il/{il}:
    get:
      tags: [Fiyat]
      summary: İl bazlı fiyat istatistiği
      security: []
      parameters:
        - { name: il, in: path, required: true, schema: { type: string } }
        - { name: kategori, in: query, schema: { type: string, default: arsa, enum: [arsa, tarla, konut, bahce, bag, zeytinlik] } }
      responses:
        "200":
          description: İl istatistik + ilçe listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  medyan: { type: number }
                  ilan_adet: { type: integer }
                  kaynak: { type: string, enum: [ilan-istatistik, ai-aggregate] }
                  ilceler:
                    type: array
                    items:
                      type: object
                      properties:
                        ilce_norm: { type: string }
                        medyan: { type: number }
                        ilan_adet: { type: integer }
  /v1/fiyat/toplu-ozet:
    get:
      tags: [Fiyat]
      summary: Tüm illerin medyan TL/m² özeti (harita choropleth için)
      description: Harita katmanı bu endpoint'ten beslenir. CDN'de 2 saat cache'lenir.
      security: []
      parameters:
        - { name: kategori, in: query, schema: { type: string, default: arsa, enum: [arsa, tarla, konut, bahce, bag, zeytinlik] } }
      responses:
        "200":
          description: Tüm iller özeti
          content:
            application/json:
              schema:
                type: object
                properties:
                  kategori: { type: string }
                  ilSayisi: { type: integer }
                  iller:
                    type: array
                    items:
                      type: object
                      properties:
                        il_norm: { type: string }
                        medyan: { type: number }
                        ilan_adet: { type: integer }
                        kaynak: { type: string, enum: [ilan, ai-baseline] }
                  guncelleme: { type: string, format: date-time }
  /v1/fiyat/toplu-ilce-ozet/{il}:
    get:
      tags: [Fiyat]
      summary: Bir ilin tüm ilçelerinin medyan özeti
      security: []
      parameters:
        - { name: il, in: path, required: true, schema: { type: string } }
        - { name: kategori, in: query, schema: { type: string, default: arsa } }
      responses:
        "200":
          description: İlçe özet listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  il: { type: string }
                  kategori: { type: string }
                  ilceler:
                    type: array
                    items:
                      type: object
                      properties:
                        ilce_norm: { type: string }
                        medyan: { type: number }
                        ilan_adet: { type: integer }
                        mahalle_sayi: { type: integer }
  /v1/fiyat/trend/{il}/{ilce}/{mahalle}:
    get:
      tags: [Fiyat]
      summary: Mahalle fiyat trendi + 6 aylık OLS projeksiyonu
      security: []
      parameters:
        - { name: il, in: path, required: true, schema: { type: string } }
        - { name: ilce, in: path, required: true, schema: { type: string } }
        - { name: mahalle, in: path, required: true, schema: { type: string } }
        - { name: kategori, in: query, schema: { type: string, default: arsa } }
      responses:
        "200":
          description: Trend verisi
          content:
            application/json:
              schema:
                type: object
                properties:
                  gecmis:
                    type: array
                    items: { $ref: '#/components/schemas/ZamanNoktasi' }
                  projeksiyon:
                    type: array
                    items:
                      type: object
                      properties:
                        yil: { type: integer }
                        ay: { type: integer }
                        tahmin: { type: number }
                        guven_alt: { type: number }
                        guven_ust: { type: number }
                  yillikDegisimYuzde: { type: number }
                  ruelDegisimYuzde: { type: number }
                  trend: { type: string, enum: [yukseliyor, dusuyor, duruyor] }
                  r2: { type: number }
                  aylikEgimTlm2: { type: number }
                  veriAyAdet: { type: integer }
                  seviye: { type: string, enum: [mahalle, ilce, il] }
        "404":
          description: Trend verisi yok

  # ── Koord Sorgu ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Kurumsal Pro tier API anahtarı.
        Format: `cdrm_<32-hex>`. /v1/api/tokens endpoint'inden oluşturulur.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token — /v1/auth/giris endpoint'inden alınır.
        /v1/api/tokens CRUD işlemleri için gereklidir.
  schemas:
    MahalleIstatistik:
      type: object
      properties:
        kaynak: { type: string, enum: [ilan-istatistik, ai-baseline] }
        medyan: { type: number, description: TL/m² medyan fiyat }
        q1: { type: number }
        q3: { type: number }
        ortalama: { type: number, nullable: true }
        ilan_adet: { type: integer }
        son_guncelleme: { type: integer, description: Unix timestamp (ms) }
        trend:
          type: array
          items: { $ref: '#/components/schemas/ZamanNoktasi' }
    ZamanNoktasi:
      type: object
      properties:
        yil: { type: integer }
        ay: { type: integer }
        medyan: { type: number }
        ilan_adet: { type: integer }
...